fix: resolve Phase 1 code review issues in DDL type and repository safety
- Change table_relation.is_active from SMALLINT to INTEGER to match Java
  entity field `Integer isActive`, preventing Hibernate type mismatch on startup
- Add safelyDeleteByDataSourceIdAndSourceIn default method to
  TableRelationRepository to guard against empty IN () SQL syntax error,
  consistent with existing safelyFindRelationsInvolvingTables pattern

+++ src/main/resources/db/init.sql
@@ -147,7 +147,7 @@ CREATE TABLE IF NOT EXISTS table_relation (
     to_column       VARCHAR(200) NOT NULL,
     source          VARCHAR(20) NOT NULL,   -- fk / naming / llm / manual
     confidence      NUMERIC(3,2) DEFAULT 1.0,
-    is_active       SMALLINT DEFAULT 1,
+    is_active       INTEGER DEFAULT 1,
     created_at      TIMESTAMP DEFAULT NOW(),
     CONSTRAINT uk_table_relation UNIQUE (data_source_id, from_table, from_column, to_table, to_column)
 );
